requirements.txt
flask==2.3.3
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
flask-cors==4.0.0
numpy==1.24.3
matplotlib==3.7.2
opencv-python==4.8.0.76
gunicorn==21.2.0